requirements.txt
base58==2.1.1
cartesapp @ git+https://github.com/prototyp3-dev/cartesapp@f46e44fa481cdfdbf5118f4737f6714bc76a94b0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cytoolz==0.12.3
eth-abi==4.2.1
eth-hash==0.6.0
eth-typing==4.0.0
eth-utils==3.0.0
exceptiongroup==1.2.0
idna==3.6
iniconfig==2.0.0
Jinja2==3.1.3
MarkupSafe==2.1.5
packaging==23.2
parsimonious==0.9.0
pillow==10.2.0
pluggy==1.4.0
pony==0.7.17
protobuf==4.25.3
py-expression-eval==0.3.14
pycryptodome==3.19.1
pydantic==1.10.14
pydantic-to-typescript==1.0.10
pytest==7.4.4
pytest-order==1.2.0
python-cartesi==0.1.1
regex==2023.12.25
requests==2.31.0
tomli==2.0.1
toolz==0.12.1
typer==0.9.0
typing_extensions==4.9.0
urllib3==2.2.1
